refactor: move the test to the correct file
