+++ src/commands/dev/config/pull.rs
@@ -156,50 +156,3 @@ pub async fn handle_config_pull(path: &Path) -> Result<bool, CliError> {
         Ok(true)
     }
 }
-
-// fn get_secret_config_files(path: &Path) -> Vec<PathBuf> {
-//     let mut overrides = OverrideBuilder::new(path);
-//     overrides.add("**/config/dev.exs").unwrap();
-//     overrides.add("**/.wukong.toml").unwrap();
-//
-//     WalkBuilder::new(path)
-//         .overrides(overrides.build().unwrap())
-//         .build()
-//         .flatten()
-//         .filter(|e| e.path().is_file())
-//         .map(|e| e.path().to_path_buf())
-//         .collect()
-// }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use assert_fs::prelude::{FileTouch, PathChild};
-
-    #[test]
-    fn test_dev_files() {
-        // three files:
-        // temp/config/dev.exs
-        // temp/config/prod.exs
-        // temp/app/config/dev.exs
-
-        let temp = assert_fs::TempDir::new().unwrap();
-        let dev_config_file = temp.child("config/dev.exs");
-        dev_config_file.touch().unwrap();
-        let prod_config_file = temp.child("config/prod.exs");
-        prod_config_file.touch().unwrap();
-        let another_dev_config_file = temp.child("app/config/dev.exs");
-        another_dev_config_file.touch().unwrap();
-
-        let files = get_secret_config_files(Some(temp.to_path_buf())).unwrap();
-        let files_names = files
-            .iter()
-            .map(|f| f.to_string_lossy())
-            .collect::<Vec<_>>();
-        assert_eq!(files.len(), 2);
-        assert!(files_names.contains(&dev_config_file.path().to_string_lossy()));
-        assert!(files_names.contains(&another_dev_config_file.path().to_string_lossy()));
-
-        temp.close().unwrap();
-    }
-}
